+++ confluence-client/src/main/kotlin/com/github/zeldigas/confclient/ConfluenceClient.kt
@@ -1,9 +1,6 @@
 package com.github.zeldigas.confclient
 
-import com.github.zeldigas.confclient.model.Attachment
-import com.github.zeldigas.confclient.model.ConfluencePage
-import com.github.zeldigas.confclient.model.PageAttachments
-import com.github.zeldigas.confclient.model.Space
+import com.github.zeldigas.confclient.model.*
 import io.ktor.http.*
 import java.nio.file.Path
 
@@ -76,6 +73,8 @@ interface ConfluenceClient {
 
     suspend fun downloadAttachment(attachment: Attachment, destination: Path)
 
+    suspend fun getUserByKey(userKey: String): User
+
 }
 
 class PageNotCreatedException(val title: String, val status: Int, val body: String?) :
@@ -87,3 +86,9 @@ class PageNotUpdatedException(val id: String, val status: Int, val body: String?
 class PageNotFoundException : RuntimeException()
 
 class TooManyPagesFound(val pages: List<ConfluencePage>) : RuntimeException()
+
+class UnknownConfluenceErrorException(val status: Int, val body: String?) :
+    RuntimeException("Unknown Confluence error: status=$status, body:\n$body")
+
+class ConfluenceApiErrorException(val status: Int, val error: String, val body: Map<String, Any?>) :
+    RuntimeException("Confluence API error: status=$error, body:\n$body")