+++ convert/src/main/kotlin/com/github/zeldigas/text2confl/convert/markdown/export/ConfluenceUserResolver.kt
@@ -1,0 +1,13 @@
+package com.github.zeldigas.text2confl.convert.markdown.export
+
+interface ConfluenceUserResolver {
+
+    companion object {
+        val NOP = object : ConfluenceUserResolver {
+            override fun resolve(userKey: String): String? = null
+        }
+    }
+
+    fun resolve(userKey: String): String?
+
+}