+++ convert/src/main/kotlin/com/github/zeldigas/text2confl/convert/markdown/export/HtmlToMarkdownConverter.kt
@@ -6,11 +6,17 @@ import com.vladsch.flexmark.util.data.MutableDataSet
 import java.nio.file.Path
 import kotlin.io.path.Path
 
-class HtmlToMarkdownConverter(resolver: ConfluenceLinksResolver, assetsLocation: String) {
+class HtmlToMarkdownConverter(
+    linkResolver: ConfluenceLinksResolver,
+    assetsLocation: String,
+    userResolver: ConfluenceUserResolver? = null
+) {
 
     companion object {
         val LINK_RESOLVER =
             DataKey<ConfluenceLinksResolver>("CONFLUENCE_LINK_RESOLVER") { ConfluenceLinksResolver.NOP }
+        val USER_RESOLVER =
+            DataKey<ConfluenceUserResolver>("CONFLUENCE_USER_RESOLVER") { ConfluenceUserResolver.NOP }
         val ASSETS_DIR = DataKey<Path>("CONFLUENCE_ASSETS_DIR") { Path("_assets") }
     }
 
@@ -18,8 +24,13 @@ class HtmlToMarkdownConverter(resolver: ConfluenceLinksResolver, assetsLocation:
         MutableDataSet()
             .set(FlexmarkHtmlConverter.SETEXT_HEADINGS, false)
             .set(FlexmarkHtmlConverter.LIST_CONTENT_INDENT, false)
-            .set(LINK_RESOLVER, resolver)
+            .set(LINK_RESOLVER, linkResolver)
             .set(ASSETS_DIR, Path(assetsLocation))
+            .also {
+                if (userResolver != null) {
+                    it.set(USER_RESOLVER, userResolver)
+                }
+            }
     )
         .htmlNodeRendererFactory { ConfluenceCustomNodeRenderer(it) }
         .build()