+++ convert/src/test/kotlin/com/github/zeldigas/text2confl/convert/asciidoc/RenderingOfLinksTest.kt
@@ -4,6 +4,7 @@ import assertk.assertThat
 import com.github.zeldigas.text2confl.convert.Attachment
 import com.github.zeldigas.text2confl.convert.PageHeader
 import com.github.zeldigas.text2confl.convert.confluence.ReferenceProvider
+import org.junit.jupiter.api.Tag
 import org.junit.jupiter.api.Test
 import kotlin.io.path.Path
 
@@ -25,7 +26,9 @@ internal class RenderingOfLinksTest : RenderingTestBase() {
             
             link:test/another.md#first-header[Link with anchor to another type via link macro]
             
-            xref:test/another.md#first-header[Link [.line-through]#with# **anchor** to `another type`]                       
+            xref:test/another.md#first-header[Link [.line-through]#with# **anchor** to `another type`]
+                                   
+            xref:a%20spaced/file.md#first-header[Link to spaced file]
         """.trimIndent(),
             referenceProvider = AsciidocReferenceProvider(
                 Path("./test.adoc"),
@@ -33,7 +36,8 @@ internal class RenderingOfLinksTest : RenderingTestBase() {
                     Path("."), mapOf(
                         Path("src.adoc") to PageHeader("Test", emptyMap()),
                         Path("another.adoc") to PageHeader("Another adoc", emptyMap()),
-                        Path("test/another.md") to PageHeader("Markdown", emptyMap())
+                        Path("test/another.md") to PageHeader("Markdown", emptyMap()),
+                        Path("a spaced/file.md") to PageHeader("File in spaced dir", emptyMap())
                     )
                 )
             )
@@ -48,6 +52,7 @@ internal class RenderingOfLinksTest : RenderingTestBase() {
             <p><ac:link ac:anchor="first-header"><ri:page ri:content-title="Markdown" ri:space-key="TEST" /><ac:plain-text-link-body><![CDATA[Link with anchor to another type]]></ac:plain-text-link-body></ac:link></p>
             <p><ac:link ac:anchor="first-header"><ri:page ri:content-title="Markdown" ri:space-key="TEST" /><ac:plain-text-link-body><![CDATA[Link with anchor to another type via link macro]]></ac:plain-text-link-body></ac:link></p>
             <p><ac:link ac:anchor="first-header"><ri:page ri:content-title="Markdown" ri:space-key="TEST" /><ac:link-body>Link <del>with</del> <strong>anchor</strong> to <code>another type</code></ac:link-body></ac:link></p>
+            <p><ac:link ac:anchor="first-header"><ri:page ri:content-title="File in spaced dir" ri:space-key="TEST" /><ac:plain-text-link-body><![CDATA[Link to spaced file]]></ac:plain-text-link-body></ac:link></p>
         """.trimIndent(),
         )
     }
@@ -165,6 +170,42 @@ internal class RenderingOfLinksTest : RenderingTestBase() {
             <p><a href="mailto:example@example.org">Send email</a></p>
         """.trimIndent(),
         )
+    }
+
+    @Test
+    @Tag("GH-136")
+    fun `Xrefstyle is used for xref rendering`() {
+        val result = toHtml(
+            """
+            = Document
+            :sectnums:
+            :xrefstyle: full
+            
+            == test
+
+            As it was described in <<another>>
+
+            == another
+            
+            section
+        """.trimIndent(),
+            referenceProvider = AsciidocReferenceProvider(
+                Path("./test.adoc"),
+                ReferenceProvider.fromDocuments(
+                    Path("."), emptyMap()
+                )
+            )
+        )
+
+        assertThat(result).isEqualToConfluenceFormat(
+            """
+            <h1>1. test<ac:structured-macro ac:name="anchor"><ac:parameter ac:name="">test</ac:parameter></ac:structured-macro></h1>
+            <p>As it was described in <ac:link ac:anchor="another"><ac:link-body>Section 2, &#8220;another&#8221;</ac:link-body></ac:link></p>
+            <h1>2. another<ac:structured-macro ac:name="anchor"><ac:parameter ac:name="">another</ac:parameter></ac:structured-macro></h1>
+            <p>section</p>
+        """.trimIndent(),
+        )
+
     }
 }