+++ convert/src/test/kotlin/com/github/zeldigas/text2confl/convert/markdown/export/HtmlToMarkdownConverterTest.kt
@@ -7,7 +7,16 @@ import org.junit.jupiter.params.provider.ValueSource
 
 class HtmlToMarkdownConverterTest {
 
-    private val converter = HtmlToMarkdownConverter(ConfluenceLinksResolver.NOP, "_assets")
+    private val converter = HtmlToMarkdownConverter(ConfluenceLinksResolver.NOP, "_assets",
+        userResolver = object : ConfluenceUserResolver {
+            override fun resolve(userKey: String): String? {
+                return when (userKey) {
+                    "known" -> "user"
+                    "known_email" -> "user@example.org"
+                    else -> null
+                }
+            }
+        })
 
     @ValueSource(
         strings = [
@@ -16,18 +25,19 @@ class HtmlToMarkdownConverterTest {
             "links",
             "tables",
             "confluence-specific",
+            "user-refs",
         ]
     )
     @ParameterizedTest
     fun `Conversion of confluence page`(pageId: String) {
-        val input = readResoource("/convert/$pageId.html")
+        val input = readResource("/convert/$pageId.html")
 
         val result = converter.convert(input)
 
-        assertThat(result).isEqualTo(readResoource("/convert/$pageId.md"))
+        assertThat(result).isEqualTo(readResource("/convert/$pageId.md"))
     }
 
-    private fun readResoource(resource: String): String {
+    private fun readResource(resource: String): String {
         return HtmlToMarkdownConverter::class.java.getResourceAsStream(resource)?.use {
             String(it.readAllBytes()).replace("\r\n", "\n")
         } ?: throw IllegalStateException("Failed to load $resource")