+++ core/src/main/kotlin/com/github/zeldigas/text2confl/core/upload/PageUploadOperations.kt
@@ -27,29 +27,45 @@ interface PageUploadOperations {
 
 sealed class PageOperationResult {
     data class NotModified(override val local: Page, override val serverPage: ServerPage) : PageOperationResult()
-    data class LocationModified(override val local: Page, override val serverPage: ServerPage, val previousParent: String, val previousTitle: String) : PageOperationResult()
+    data class LocationModified(
+        override val local: Page,
+        override val serverPage: ServerPage,
+        val previousParent: String,
+        val previousTitle: String
+    ) : PageOperationResult()
+
     data class Created(override val local: Page, override val serverPage: ServerPage) : PageOperationResult()
-    data class ContentModified(override val local: Page, override val serverPage: ServerPage, val parentChanged: Boolean = false) : PageOperationResult()
+    data class ContentModified(
+        override val local: Page,
+        override val serverPage: ServerPage,
+        val parentChanged: Boolean = false
+    ) : PageOperationResult()
 
     abstract val local: Page
     abstract val serverPage: ServerPage
 }
 
 sealed class LabelsUpdateResult {
-    data class Updated(val added: List<String>, val removed: List<String>): LabelsUpdateResult()
-    object NotChanged: LabelsUpdateResult()
+    data class Updated(val added: List<String> = emptyList(), val removed: List<String> = emptyList()) :
+        LabelsUpdateResult()
+
+    object NotChanged : LabelsUpdateResult()
 }
 
 sealed class AttachmentsUpdateResult {
-    data class Updated(val added: List<com.github.zeldigas.text2confl.convert.Attachment>,
-                  val modified:List<com.github.zeldigas.text2confl.convert.Attachment>,
-                  val removed: List<Attachment>): AttachmentsUpdateResult()
-    object NotChanged: AttachmentsUpdateResult()
+    data class Updated(
+        val added: List<com.github.zeldigas.text2confl.convert.Attachment> = emptyList(),
+        val modified: List<com.github.zeldigas.text2confl.convert.Attachment> = emptyList(),
+        val removed: List<Attachment> = emptyList()
+    ) : AttachmentsUpdateResult()
+
+    object NotChanged : AttachmentsUpdateResult()
 }
 
 abstract class PageOperationException(message: String, cause: Exception? = null) : RuntimeException(message, cause)
 
-data class PageNotFoundException(val space: String, val title: String): PageOperationException("Page $title in space $space not found")
+data class PageNotFoundException(val space: String, val title: String) :
+    PageOperationException("Page $title in space $space not found")
 
 enum class ChangeDetector(
     val extraData: Set<String>,